test(config): split suite by subsystem
